Adding contract size output to compile task for buidler
